Add portfolio cost calculation in pcost.py
Implemented a function called 'portfolio_cost' in 'pcost.py' that calculates the total cost of portfolio based on provided data. It takes a 'filename' as an argument, reads the respective file and computes the total_cost by taking shares and cost into consideration. It also takes care of incorrect values by using try and exception.

Completion of the Introduction section of the Python course