Implement cost calculation and stock selling in Stock class
Enhanced the Stock class in 'stock.py' by adding the cost property which calculates the cost depending on the current shares and price, and a 'sell' method that adjusts the number of shares. Refactored the imports from 'typedproperty.py' to be more efficient. Also, created a 'test_stock.py' for unit testing of the modifications made in the Stock class.